requirements.txt
absl-py==2.1.0
accelerate==1.2.1
annotated-types==0.7.0
anyio==4.7.0
beautifulsoup4==4.13.4
bert-score==0.3.13
cfgv==3.4.0
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
datasets==3.2.0
dill==0.3.8
distlib==0.3.9
distro==1.9.0
docker-pycreds==0.4.0
einops==0.8.0
entmax==1.3
et-xmlfile==2.0.0
evaluate==0.4.3
execnet==2.1.1
executorch==0.4.0
expecttest==0.3.0
flatbuffers==24.12.23
gdown==5.2.0
gitdb==4.0.12
gitpython==3.1.44
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.0
humanfriendly==10.0
hypothesis==6.123.4
identify==2.6.5
info-nce-pytorch==0.1.4
iniconfig==2.0.0
inquirerpy==0.3.4
jiter==0.8.2
lxml==5.3.0
markdown==3.7
multiprocess==0.70.16
nltk==3.9.1
nodeenv==1.9.1
numpy==1.23.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.59.4
openpyxl==3.1.5
optimum==1.23.3
parameterized==0.9.0
pfzy==0.3.4
platformdirs==4.3.6
pluggy==1.5.0
portalocker==3.1.1
pre-commit==4.0.1
prompt-toolkit==3.0.48
protobuf==5.29.2
ptflops==0.7.4
pyarrow==18.1.0
pybind11==2.13.6
pydantic==2.10.4
pydantic-core==2.27.2
pylcs==0.1.1
pytest==8.3.4
pytest-xdist==3.6.1
regex==2024.11.6
rouge-score==0.1.2
ruamel-yaml==0.18.10
ruamel-yaml-clib==0.2.12
sacrebleu==2.5.1
safetensors==0.4.5
sentence-transformers==3.3.1
sentencepiece==0.2.0
sentry-sdk==2.22.0
setproctitle==1.3.5
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
sparsemax==0.1.9
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tokenizers==0.21.0
torch==2.5.0
torchaudio==2.5.0
torchvision==0.20.0
transformers==4.47.1
triton==3.1.0
typing-extensions==4.12.2
vec2text==0.0.12
virtualenv==20.28.1
wandb==0.19.8
wcwidth==0.2.13
werkzeug==3.1.3
xlrd==2.0.2
xxhash==3.5.0